secondary flame, touchups
