--- secrets.yaml
@@ -1,8 +1,0 @@
-wifi_ssid: "myssidname"
-wifi_password: "mywifipassword"
-ap_password: "updatepassword"
-api_encryption_key: "homeassistant-secret"
-# Generate a secret here https://esphome.io/components/api/
-web_password: "webpassword"
-
-
